requirements.txt
google-adk
google-cloud-aiplatform
faker
streamlit
